Move analytics provider files
