+++ src/palace/manager/service/analytics/local.py
